# src/Modules/Catalog/Catalog-OpenAPI.yaml
openapi: 3.0.0
info:
  title: Catalog API
  version: v1
servers:
  - url: https://localhost:5001
paths:
...
  /api/catalog/brands:
    get:
      tags:
        - catalog/brands
      operationId: Catalog_Brand-GetAll
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            format: int32
          x-position: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
          x-position: 2
        - name: searchString
          in: query
          schema:
            type: string
            nullable: true
          x-position: 3
        - name: orderBy
          in: query
          schema:
            type: string
            nullable: true
          x-position: 4
        - name: includeDeleted
          in: query
          schema:
            type: boolean
            default: false
          x-position: 5
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedResultOfBrandsResponseModel'
        401:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        404:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        500:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
      security:
        - JWT: []
    post:
      tags:
        - catalog/brands
      operationId: Catalog_Brand-Post
      requestBody:
        x-name: model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandModel'
        required: true
        x-position: 1
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultOfEntityCreatedResponseModel'
        400:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetailsModel'
        500:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
      security:
        - JWT: []
...
  /api/catalog/brands/export:
    get:
      tags:
        - catalog/brands
      operationId: Catalog_Brand-Export
      parameters:
        - name: searchString
          in: query
          schema:
            type: string
            default: ''
            nullable: true
          x-position: 1
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultOfStringResponseModel'
        401:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        404:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        500:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
      security:
        - JWT: []
  /api/catalog/brands/import:
    post:
      tags:
        - catalog/brands
      operationId: Catalog_Brand-Import
      requestBody:
        x-name: model
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequestModel'
        required: true
        x-position: 1
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultOfIntegerResponseModel'
        401:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        404:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
        500:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetailsModel'
      security:
        - JWT: []

  # Products ======================================================
...
components:
  schemas:
    ResultResponseModel:
      type: object
      additionalProperties: false
      required:
        - messages
        - isSuccess
      properties:
        messages:
          type: array
          nullable: true
          items:
            type: string
        isSuccess:
          type: boolean
    ResultOfStringResponseModel:
      allOf:
        - $ref: '#/components/schemas/ResultResponseModel'
        - type: object
          additionalProperties: false
          properties:
            value:
              type: string
              nullable: true
    ResultOfIntegerResponseModel:
      allOf:
        - $ref: '#/components/schemas/ResultResponseModel'
        - type: object
          additionalProperties: false
          properties:
            value:
              type: integer
              format: int32
    ProblemDetailsModel:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
    ValidationProblemDetailsModel:
      allOf:
        - $ref: '#/components/schemas/ProblemDetailsModel'
        - type: object
          additionalProperties: false
          required:
            - errors
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    ResultOfEntityCreatedResponseModel:
      allOf:
        - $ref: '#/components/schemas/ResultResponseModel'
        - type: object
          additionalProperties: false
          properties:
            entityId:
              type: string
...
    PagedResultOfBrandsResponseModel:
      allOf:
        - $ref: '#/components/schemas/ResultOfBrandsResponseModel'
        - type: object
          additionalProperties: false
          properties:
            currentPage:
              type: integer
              format: int32
            totalPages:
              type: integer
              format: int32
            totalCount:
              type: integer
              format: int64
            pageSize:
              type: integer
              format: int32
            hasPreviousPage:
              type: boolean
            hasNextPage:
              type: boolean
...
    ResultOfBrandsResponseModel:
      allOf:
        - $ref: '#/components/schemas/ResultResponseModel'
        - type: object
          additionalProperties: false
          properties:
            value:
              type: array
              nullable: true
              items:
                $ref: '#/components/schemas/BrandModel'
    BrandModel:
      allOf:
        - type: object
          additionalProperties: false
          properties:
            id:
              type: string
              format: guid
            name:
              type: string
              nullable: true
            description:
              type: string
              nullable: true
            keywords:
              type: array
              nullable: true
              items:
                type: string
            pictureSvg:
              type: string
              nullable: true
            pictureFileName:
              type: string
              nullable: true
            pictureUri:
              type: string
              nullable: true
            auditState:
              nullable: true
              oneOf:
                - $ref: '#/components/schemas/AuditStateModel'
    # Produts ======================================================
...
    AuditStateModel:
      type: object
      additionalProperties: false
      properties:
        createdBy:
          type: string
          nullable: true
        createdDate:
          type: string
          format: date-time
        createdDescription:
          type: string
          nullable: true
        updatedBy:
          type: string
          nullable: true
        updatedDate:
          type: string
          format: date-time
          nullable: true
        updatedDescription:
          type: string
          nullable: true
        updatedReasons:
          type: array
          nullable: true
          items:
            type: string
        deactivated:
          type: boolean
          nullable: true
        deactivatedReasons:
          type: array
          nullable: true
          items:
            type: string
        deactivatedBy:
          type: string
          nullable: true
        deactivatedDate:
          type: string
          format: date-time
          nullable: true
        deactivatedDescription:
          type: string
          nullable: true
        deleted:
          type: boolean
          nullable: true
        deletedBy:
          type: string
          nullable: true
        deletedDate:
          type: string
          format: date-time
          nullable: true
        deletedReason:
          type: string
          nullable: true
        deletedDescription:
          type: string
          nullable: true
        lastActionDate:
          type: string
          format: date-time
          nullable: true
    UploadRequestModel:
      type: object
      additionalProperties: false
      properties:
        fileName:
          type: string
          nullable: true
        extension:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        data:
          type: string
          format: byte
          nullable: true
  securitySchemes:
    JWT:
      type: apiKey
      description: Input your JWT token in this format - Bearer {your token here} to access this API
      name: Authorization
      in: header
      # bearerFormat: JWT
security:
  - JWT: []